% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kifidi-package.R
\docType{package}
\name{Kifidi-package}
\alias{Kifidi}
\alias{Kifidi-package}
\title{Kifidi: Tools for summarizing and analyzing grouped environmental data}
\description{
The Kifidi package provides tools for summarizing and visualizing grouped numerical data,
especially for environmental and ecological datasets. It includes functions for generating
statistical summaries, plotting means with error bars, performing grouped regression analysis,
and generating frequency counts by group.
}
\section{Main Functions}{

\describe{
\item{\code{\link{summarize_data}}}{Provides statistical summaries (mean, SD, N, etc.) of numeric data grouped by one or two categorical variables.}
\item{\code{\link{plot_means}}}{Creates bar plots of means with optional error bars.}
\item{\code{\link{counts}}}{Generates frequency tables or counts of observations by grouping variables.}
\item{\code{\link{plot_group_regressions}}}{Performs and plots linear regressions grouped by a factor variable.}
\item{\code{\link{plot_lmm_regressions}}}{Plots group-level and fixed-effect regression lines from a linear mixed-effects model.}
}
}

\author{
Oswald Omuron
}
\keyword{package}
