\name{Kifidi-package}
\alias{Kifidi}
\docType{package}
\title{Kifidi package by Oswald Omuron}
\description{
  Kifidi v.0.1.0
}
\details{
  Package: Kifidi \cr
  Type: Package \cr
  Title: Kifidi v.0.1.0 \cr
  Version: 0.1.0 \cr
  Author: Oswald Omuron \cr
  Maintainer: Oswald Omuron <oswaldomuron@gmail.com> \cr
  Description: An overview of how to use the package, including the most important functions under See Also below. \cr
  License: GPL-3 \cr
}
\author{
  Oswald Omuron

  Maintainer: Oswald Omuron <oswaldomuron@gmail.com>
}
\references{
  \url{https://github.com/OswaldOmuron/Kifidi}
}
\seealso{
  Optional links to other man pages, e.g. \code{\link{summarize_data}},
  \code{\link{plot_means}}
}


\keyword{package}

